Lots of issues with Python indentation
Switched from looking for a TokenList in the parentage to looking specifically for a Python_MultilineStatement. But I also had to changes all the 'else' and 'elif' clauses to get a customized start-of-line parser because the container is not yet complete so can't really look at the _startingChar for the 'if' clause. Have to look for an AbstractStatement which seems to work.