New Transformation system is starting to work.
No output at all yet. Need to implement mid() and len() functions. But lots of machinery is in place and seems to be working. This approach puts all the logic with the parse and interpret logic, rather than a separate package. Also, it uses AbstractStatement etc. instead of generics. Generics were slightly better but much harder to work with. Could possibly add generics back in later.